Move rustc_middle::middle::stability to rustc_crate.
