+++ src/librustdoc/html/render/mod.rs
@@ -50,12 +50,12 @@ use std::sync::Arc;
 use itertools::Itertools;
 use rustc_ast_pretty::pprust;
 use rustc_attr::StabilityLevel;
+use rustc_crate::stability;
 use rustc_data_structures::flock;
 use rustc_data_structures::fx::{FxHashMap, FxHashSet};
 use rustc_hir as hir;
 use rustc_hir::def_id::{DefId, LOCAL_CRATE};
 use rustc_hir::Mutability;
-use rustc_middle::middle::stability;
 use rustc_span::edition::Edition;
 use rustc_span::hygiene::MacroKind;
 use rustc_span::source_map::FileName;
